functions/requirements.txt
functions-framework==3.*
Pillow>=10.0.0
